requirements.txt
streamlit==1.25.0
pypdf==3.17.2
openai==0.27.2
langchain==0.0.200
ollama  # Ollama integration (installed via script, so placeholder)
dvc==3.24.0
numpy==1.24.3
pandas==2.0.3
PyYAML==6.0.1  # For reading config.yaml
pinecone-client  # Pinecone for vector databases
python-dotenv
-e .

